pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deer-flow"
version = "0.1.0"
description = "DeerFlow project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "httpx>=0.28.1",
    "langchain-community>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "readabilipy>=0.3.0",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sse-starlette>=1.6.5",
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "yfinance>=0.2.54",
    "litellm>=1.63.11",
    "json-repair>=0.7.0",
    "jinja2>=3.1.3",
    "duckduckgo-search>=8.0.0",
    "inquirerpy>=0.3.4",
    "arxiv>=2.2.0",
    "mcp>=1.11.0",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-deepseek>=0.1.3",
    "langchain-google-genai>=2.0.6",
    "wikipedia>=1.4.0",
    "langchain-tavily<0.3",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "langgraph-checkpoint-postgres==2.0.21",
    "pymilvus>=2.3.0",
    "langchain-milvus>=0.2.1",
    "psycopg[binary]>=3.2.9",
    "rdkit>=2025.9.1",
    "joblib>=1.5.2",
    "addict>=2.4.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "numba>=0.57.0",
    # Workflow API dependencies
    "flask>=3.1.2",
    "flask-compress>=1.17,<1.18",
    "flask-cors>=6.0.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-orjson>=2.0.0",
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "sqlalchemy>=2.0.29",
    "celery>=5.5.2",
    "redis[hiredis]>=6.1.0",
    "psycopg2-binary>=2.9.6",
    "psycogreen>=1.0.2",
    "gevent>=25.9.1",
    "pydantic>=2.11.4",
    "pydantic-extra-types>=2.10.3",
    "pydantic-settings>=2.11.0",
    "beautifulsoup4>=4.12.2",
    "bs4>=0.0.1",
    "chardet>=5.1.0",
    "jieba>=0.42.1",
    "markdown>=3.5.1",
    "openpyxl>=3.1.5",
    "python-docx>=1.1.0",
    "pyyaml>=6.0.1",
    "pyjwt>=2.10.1",
    "pycryptodome>=3.19.1",
    "tiktoken>=0.9.0",
    "transformers>=4.56.1",
    "unstructured[docx,epub,md,ppt,pptx]>=0.16.1",
    "pypdfium2>=4.30.0",
    "webvtt-py>=0.5.1",
    "sseclient-py>=1.8.0",
    "httpx-sse>=0.4.0",
    "sendgrid>=6.12.3",
    "resend>=2.9.0",
    "croniter>=6.0.0",
    "apscheduler>=3.11.0",
    "weaviate-client>=4.17.0",
    "weave>=0.52.16",
    "langfuse>=2.51.3",
    "langsmith>=0.1.77",
    "sentry-sdk[flask]>=2.28.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-distro>=0.48b0",
    "opentelemetry-instrumentation>=0.48b0",
    "opentelemetry-instrumentation-flask>=0.48b0",
    "opentelemetry-instrumentation-celery>=0.48b0",
    "deprecated>=1.2.14",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-redis>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-proto>=1.27.0",
    "boto3>=1.35.99",
    "google-cloud-aiplatform>=1.49.0",
    "azure-identity>=1.16.1",
    "mlflow-skinny>=3.0.0",
    "gmpy2>=2.2.1",
    "cachetools>=5.3.0",
    "packaging>=23.2",
    "yarl>=1.18.3",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "langgraph-cli[inmem]>=0.2.10",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "asyncpg-stubs>=0.30.2",
    "mongomock>=4.3.0",
    "pytest-postgresql>=7.0.2",
]

[tool.uv]
required-version = ">=0.6.15"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.report]
fail_under = 25

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
extend-include = ["*.pyi"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
exclude = ['^/build/']

